[Refactor] Move graphqloptions to its own class (#6496)
